Fix getTicketsById (the big endpoint that uses Kafka)

+++ src/main/kotlin/it/group24/lab5/webapp2/ticketcatalogue/dtos/OrderDTO.kt
@@ -6,7 +6,7 @@ data class OrderDTO(
     var id: Long?,
     var status: String?,
     var ticketId: Long?,
-    var quantity: Long?,
+    var quantity: Int?,
     var userId: Long?
 )
